Object-orientified the Matrix class
Matrix class no longer deals directly with double array inputs, but now
with Matrix objects.

A new matrix object can be created with new Matrix(double array)
As such, a few functions are no longer static and there is no need for
separate functions for integer matricies as everything is converted to
doubles. In a future update I will enhance the print function to not
print the .0 when the value is essentially an integer and allow from
creating of either a integer or double double-array from a matrix
object.