Further improvements to determinant
fixed for some weird cases where it would not work. The new method first
checks if the last row has any zeros in it, and if it does, sums up all
the rows and makes that the new last row. This ensures that the matrix
is not singular, will not cause a divide by zero error, and replaces the
two separate if statements that were supposed to do that before but
weren't working properly.